+++ changelog.md
@@ -7,11 +7,12 @@
 ### Added
 
 - Added support for new setter plugin APIs, specifically: `@proxy`, `@shared`, `@static`, `@tables`, `@tables-indexes`, `@views`
-- Allow developers to opt into enabling setter plugin Lambdas to be overridden by userland Architect project manifest with a `_override` flag; fixes #1352
 
 
 ### Changed
 
+- Lambdas defined in the userland Architect project manifest now override conflicting Lambdas returned by plugins (instead of throwing validation errors); fixes #1352
+  - Plugin developers can now use a `required` flag to enforce a validation error should their plugin conflict with userland Lambdas
 - `@tables` and `@tables-indexes` can now accept lower case key types (e.g. `*string` instead of `*String`)
 - `@tables` and `@tables-indexes` can also accept `*` and `**` as a shortcut for string-type primary and sort keys
 - Changed plugin function property tags from `plugin|type` to `_plugin|_type` to indicate internal property namespacing